Card display / dealing
Made a Game Over! Screen. Still needs implementation to display when the player loses, but has buttons that restart the game and return to the game suite title screen.

Made 'Hit' method and placeholders for Bet() and Stand()

Made CardBrain script, which is to be put on cards. Can configure whether it is a player card or dealer card. What this does is it changes the card's sprite based on the index of the hand it is a part of. The hands are Card[] in the GameManager object